Display obfuscated word without the consonants.
The program is in a very rudimentary state. A function
obfuscate_word has been added to main.c to replace the consonants
in the word (which is used for playing hangman) with hyphens before
displaying the word in that form with hyphens and vowels. The
player will be expected to guess the missing letters.

As of now, the program only does the obfuscation for one word,
that is: 'hangman'.