Fix vowel guess not handled correctly.
Previously, the program used to print a message asking the user to
only type consonants into the guess prompt. But the check_guess
function wouldn't return to the guess_handler function after that.
Instead, the entered vowel would be matched with an existing vowel
in the word. The program would, thus, erroneously come to the
conclusion that a letter had been guessed correctly when in fact,
vowels are not meant to be guessed and are displayed to the user
from the very beginning.