Match char is moved to separate function.
Earlier, if a valid character was entered into the prompt, the
program would start checking to see if the char existed in the
word. This code was a part of the check_guess function and the
number of levels of indentation had become unmanageable. So, the
code was moved to a separate function match_char.